pyproject.toml
[tool.poetry]
name = "receiptly"
version = "0.1.0"
description = ""
authors = ["Samyar <samyar1380@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
requests = "^2.32.3"
django = "^5.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
graphene-django = "^3.2.2"
strawberry-graphql-django = "^0.48.0"
django-choices-field = "^2.3.0"
redis = {extras = ["hiredis"], version = "^5.1.1"}
strawberry-django-auth = "^0.378.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
